requirements.txt
Django==2.1.8
phonenumbers
django-phonenumber-field
us
